02/14 front: api 요청하는 함수들 url 수정
back:
url 매핑을 위해 기존의 JSONcontroller 삭제 후 RestController 로 수정

+++ src/main/java/techsuppDev/techsupp/domain/Product.java
@@ -14,7 +14,7 @@ import javax.persistence.*;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long ID;
+    int ID;
     String picture;
     String product;
     String investment;
